More on docs and solved bug in read_bsa Import oplot explicitly IO: One .jl file per extension
