+++ src/miscellaneous/helper.jl
@@ -1,17 +1,16 @@
 @doc doc"""
 Concatanate `strings` with a `separator` between each.
 
-### Input
+### Arguments
 
-* strings: Array of strings to place one after another
-* separator: String to place between each string (Default: ` `)
+* `strings`: Array of strings to place one after another
+* `separator`: String to place between each string (Default: ` `)
 
 ### Output
 
 String consisting of all input strings
 """ ->
 function append_strings(strings::Union(Array{String}, Array{ASCIIString}); separator::String=" ")
-
     newString = strings[1]
     if length(strings) > 1
         for n = 2:length(strings)
@@ -29,16 +28,16 @@ append_strings(string::String) = string
 Return a new processing key with the number incremented.
 It checks for existing keys and returns a string with the next key to be used.
 
-### Input
+### Arguments
 
-* d: Dictionary containing existing keys
-* key_name: Base of the
+* `d`: Dictionary containing existing keys
+* `key_name`: Base of the
 
-### Output
+### Returns
 
 * String with new key name
 
-### Example
+### Returns
 
 ```julia
 results_storage = Dict()
@@ -51,7 +50,6 @@ results_storage[new_processing_key(results_storage, "FTest")] = 49
 ```
 """ ->
 function new_processing_key(d::Dict, key_name::String)
-
     key_numb = 1
     key = string(key_name, key_numb)
     while haskey(d, key)
@@ -65,16 +63,16 @@ end
 @doc doc"""
 Find dictionary keys containing a string.
 
-### Input
+### Arguments
 
-* d: Dictionary containing existing keys
-* partial_key: String you want to find in key names
+* `d`: Dictionary containing existing keys
+* `partial_key`: String you want to find in key names
 
-### Output
+### Returns
 
 * Array containg the indices of dictionary containing the partial_key
 
-### Example
+### Returns
 
 ```julia
 results_storage = Dict()
@@ -90,7 +88,6 @@ find_keys_containing(results_storage, "FTest")
 ```
 """ ->
 function find_keys_containing(d::Dict, partial_key::String)
-
     valid_keys = [beginswith(i, partial_key) for i = collect(keys(d))]
     findin(valid_keys, true)
 end
@@ -99,15 +96,15 @@ end
 @doc doc"""
 Extract the path, filename and extension of a file
 
-### Input
+### Arguments
 
-* fname: String with the full path to a file
+* `fname`: String with the full path to a file
 
 ### Output
 
 * Three strings containing the path, file name and file extension
 
-### Example
+### Returns
 
 ```julia
 fileparts("/Users/test/subdir/test-file.bdf")
@@ -116,20 +113,15 @@ fileparts("/Users/test/subdir/test-file.bdf")
 ```
 """ ->
 function fileparts(fname::String)
-
     if fname==""
-        pathname = ""
-        filename = ""
+        pathname  = ""
+        filename  = ""
         extension = ""
     else
-
         separators = sort(unique([search(fname, '/', i) for i = 1:length(fname)]))
         pathname = fname[1:last(separators)]
-
         extension  = last(sort(unique([search(fname, '.', i) for i = 1:length(fname)])))
-
         filename = fname[last(separators)+1:extension-1]
-
         extension  = fname[extension+1:end]
     end
 
@@ -140,16 +132,16 @@ end
 @doc doc"""
 Find the closest number to a target in an array and return the index
 
-### Input
+### Arguments
 
-* list: Array containing numbers
-* target: Number to find closest to in the list
+* `list`: Array containing numbers
+* `target`: Number to find closest to in the list
 
 ### Output
 
 * Index of the closest number to the target
 
-### Example
+### Returns
 
 ```julia
 _find_closest_number_idx([1, 2, 2.7, 3.2, 4, 3.1, 7], 3)
@@ -158,7 +150,6 @@ _find_closest_number_idx([1, 2, 2.7, 3.2, 4, 3.1, 7], 3)
 ```
 """ ->
 function _find_closest_number_idx{T <: Number}(list::Array{T, 1}, target::Number)
-
     diff_array = abs(list .- target)
     targetIdx  = findfirst(diff_array , minimum(diff_array))
 end
