+++ src/types/SSR/Statistics.jl
@@ -11,13 +11,13 @@ bin across epochs.
 
 ### Input
 
-* s: Steady state response type
-* freq_of_interest: frequency to analyse (modulation rate)
-* ID: value to store as ID (" ")
-* data_type: what to run the fft on (epochs)
-* fs: sampling rate (SSR sampling rate)
-* num_resample: number of bootstrapping interations to make (1000)
-* results_key: Where in the processing dictionary to store results ("statistics")
+* `s`: Steady state response type
+* `freq_of_interest`: frequency to analyse (modulation rate)
+* `ID`: value to store as ID (" ")
+* `data_type`: what to run the fft on (epochs)
+* `fs`: sampling rate (SSR sampling rate)
+* `num_resample`: number of bootstrapping interations to make (1000)
+* `results_key`: Where in the processing dictionary to store results ("statistics")
 
 ### Output
 
@@ -27,7 +27,6 @@ bin across epochs.
 
 ```julia
 s = bootstrap(s, N=100)
-
 ```
 """ ->
 function bootstrap(s::SSR; freq_of_interest::Union(Real, AbstractArray) = modulationrate(s), ID::String = "",
